Rename 'statistics' package to 'stats' to avoid name conflict
