+++ api/stats/db_queries/node_availability.py
