+++ tests/stats/test_model_layer.py
@@ -1,5 +1,5 @@
 from tests.test_case import TestCase
-from api.statistics.model_layer import get_sessions_country_stats
+from api.stats.model_layer import get_sessions_country_stats
 from datetime import datetime, timedelta
 from models import db, Session
